+++ docs/docs.go
@@ -2129,10 +2129,60 @@ const docTemplate = `{
                     "200": {
                         "description": "OK",
                         "schema": {
-                            "type": "array",
-                            "items": {
-                                "$ref": "#/definitions/rest.getPostsMigrationResponseData"
-                            }
+                            "$ref": "#/definitions/getPostsMigrationResponseData"
+                        }
+                    }
+                }
+            },
+            "put": {
+                "security": [
+                    {
+                        "AppUserAuth": []
+                    }
+                ],
+                "description": "Sets a configuration flag that the posts migration process is successfull and all related APIs to be redirected to Social BB",
+                "tags": [
+                    "BBS"
+                ],
+                "operationId": "BBSPutPostsMigrationFlag",
+                "parameters": [
+                    {
+                        "description": "body data",
+                        "name": "data",
+                        "in": "body",
+                        "required": true,
+                        "schema": {
+                            "$ref": "#/definitions/postsMigrationFlagData"
+                        }
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/postsMigrationFlagData"
+                        }
+                    }
+                }
+            }
+        },
+        "/api/bbs/posts-migration-flag": {
+            "get": {
+                "security": [
+                    {
+                        "AppUserAuth": []
+                    }
+                ],
+                "description": "Gets all groups and all posts without any restrictions",
+                "tags": [
+                    "BBS"
+                ],
+                "operationId": "BBSGetPostsMigrationFlag",
+                "responses": {
+                    "200": {
+                        "description": "OK",
+                        "schema": {
+                            "$ref": "#/definitions/postsMigrationFlagData"
                         }
                     }
                 }
@@ -4792,14 +4842,36 @@ const docTemplate = `{
                 }
             }
         },
-        "/group/{group-id}/members": {
-            "put": {
+        "/authman/synchronize": {
+            "post": {
                 "security": [
                     {
                         "AppUserAuth": []
                     }
                 ],
-                "description": "Updates a membership. Only the status can be changed.",
+                "description": "Synchronizes Authman groups membership",
+                "consumes": [
+                    "application/json"
+                ],
+                "tags": [
+                    "Client"
+                ],
+                "operationId": "InternalSynchronizeAuthman",
+                "responses": {
+                    "200": {
+                        "description": "OK"
+                    }
+                }
+            }
+        },
+        "/group/{group-id}/members": {
+            "post": {
+                "security": [
+                    {
+                        "AppUserAuth": []
+                    }
+                ],
+                "description": "Create multiple members in group with desired status",
                 "consumes": [
                     "application/json"
                 ],
@@ -4809,7 +4881,7 @@ const docTemplate = `{
                 "tags": [
                     "Admin"
                 ],
-                "operationId": "AdminCreateMemberships",
+                "operationId": "MultiCreateMembers",
                 "parameters": [
                     {
                         "type": "string",
@@ -4824,10 +4896,7 @@ const docTemplate = `{
                         "in": "body",
                         "required": true,
                         "schema": {
-                            "type": "array",
-                            "items": {
-                                "$ref": "#/definitions/model.MembershipStatus"
-                            }
+                            "$ref": "#/definitions/rest.createMembershipsRequest"
                         }
                     },
                     {
@@ -4968,17 +5037,6 @@ const docTemplate = `{
                     "items": {
                         "$ref": "#/definitions/ToMember"
                     }
-                }
-            }
-        },
-        "EventResponse": {
-            "type": "object",
-            "properties": {
-                "event_id": {
-                    "type": "string"
-                },
-                "user_id": {
-                    "type": "string"
                 }
             }
         },
@@ -5234,25 +5292,6 @@ const docTemplate = `{
                     "type": "string"
                 },
                 "user_id": {
-                    "type": "string"
-                }
-            }
-        },
-        "GroupMembershipResponse": {
-            "type": "object",
-            "properties": {
-                "id": {
-                    "type": "string"
-                },
-                "user_id": {
-                    "type": "string"
-                }
-            }
-        },
-        "GroupResponse": {
-            "type": "object",
-            "properties": {
-                "id": {
                     "type": "string"
                 }
             }
@@ -5657,17 +5696,6 @@ const docTemplate = `{
                 }
             }
         },
-        "PostResponse": {
-            "type": "object",
-            "properties": {
-                "id": {
-                    "type": "string"
-                },
-                "user_id": {
-                    "type": "string"
-                }
-            }
-        },
         "Sender": {
             "type": "object",
             "properties": {
@@ -5778,6 +5806,9 @@ const docTemplate = `{
                 },
                 "image_url": {
                     "type": "string"
+                },
+                "members": {
+                    "$ref": "#/definitions/model.DefaultMembershipConfig"
                 },
                 "membership_questions": {
                     "type": "array",
@@ -6014,6 +6045,23 @@ const docTemplate = `{
                 }
             }
         },
+        "getPostsMigrationResponseData": {
+            "type": "object",
+            "properties": {
+                "groups": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/Group"
+                    }
+                },
+                "posts": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/model.Post"
+                    }
+                }
+            }
+        },
         "getPutAdminGroupIDsForEventIDRequestAndResponse": {
             "type": "object",
             "properties": {
@@ -6176,6 +6224,29 @@ const docTemplate = `{
                 }
             }
         },
+        "model.DefaultMembershipConfig": {
+            "type": "object",
+            "required": [
+                "status"
+            ],
+            "properties": {
+                "net_ids": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                },
+                "status": {
+                    "type": "string",
+                    "enum": [
+                        "admin",
+                        "member",
+                        "pending",
+                        "rejected"
+                    ]
+                }
+            }
+        },
         "model.MembershipStatus": {
             "type": "object",
             "properties": {
@@ -6296,26 +6367,34 @@ const docTemplate = `{
                 "events": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/EventResponse"
+                        "$ref": "#/definitions/Event"
                     }
                 },
                 "group_memberships": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/GroupMembershipResponse"
+                        "$ref": "#/definitions/GroupMembership"
                     }
                 },
                 "groups": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/GroupResponse"
+                        "$ref": "#/definitions/Group"
                     }
                 },
                 "posts": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/PostResponse"
+                        "$ref": "#/definitions/model.Post"
                     }
+                }
+            }
+        },
+        "postsMigrationFlagData": {
+            "type": "object",
+            "properties": {
+                "migrated": {
+                    "type": "boolean"
                 }
             }
         },
@@ -6464,19 +6543,13 @@ const docTemplate = `{
                 }
             }
         },
-        "rest.getPostsMigrationResponseData": {
+        "rest.createMembershipsRequest": {
             "type": "object",
             "properties": {
-                "groups": {
+                "members": {
                     "type": "array",
                     "items": {
-                        "$ref": "#/definitions/Group"
-                    }
-                },
-                "posts": {
-                    "type": "array",
-                    "items": {
-                        "$ref": "#/definitions/model.Post"
+                        "$ref": "#/definitions/model.MembershipStatus"
                     }
                 }
             }